Move IPC into meteor directory
