+++ npm-packages/meteor-vite/src/bin/worker/vite-server.ts
@@ -2,12 +2,12 @@ import FS from 'fs/promises';
 import Path from 'path';
 import { createServer, resolveConfig, ViteDevServer } from 'vite';
 import Logger from '../../Logger';
-import MeteorEvents, { MeteorIPCMessage } from './IPC/MeteorEvents';
+import MeteorEvents, { MeteorIPCMessage } from '../../meteor/IPC/MeteorEvents';
 import { MeteorViteConfig } from '../../vite/MeteorViteConfig';
 import { MeteorStubs } from '../../vite';
 import { ProjectJson } from '../../vite/plugin/MeteorStubs';
 import { RefreshNeeded } from '../../vite/ViteLoadRequest';
-import CreateIPCInterface, { IPCReply } from './IPC/interface';
+import CreateIPCInterface, { IPCReply } from '../../meteor/IPC/interface';
 
 let server: ViteDevServer & { config: MeteorViteConfig };
 let viteConfig: MeteorViteConfig;
